+++ forms-flow-bpm/src/main/java/org/camunda/bpm/extension/hooks/listeners/FormSubmissionListener.java
@@ -9,6 +9,8 @@ import org.camunda.bpm.extension.hooks.services.FormSubmissionService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
+
+import java.io.IOException;
 import java.util.logging.Logger;
 
 /**
@@ -17,7 +19,7 @@ import java.util.logging.Logger;
  * @author sumathi.thirumani@aot-technologies.com
  */
 @Component
-public class FormSubmissionListener implements ExecutionListener, TaskListener {
+public class FormSubmissionListener extends BaseListener implements ExecutionListener, TaskListener {
 
     private final Logger LOGGER = Logger.getLogger(FormSubmissionListener.class.getName());
 
@@ -26,15 +28,24 @@ public class FormSubmissionListener implements ExecutionListener, TaskListener {
 
     @Override
     public void notify(DelegateExecution execution) {
-        createRevision(execution);
+        try {
+            createRevision(execution);
+        } catch (IOException e) {
+            handleException(execution, ExceptionSource.EXECUTION, e);
+        }
     }
 
     @Override
     public void notify(DelegateTask delegateTask) {
-        createRevision(delegateTask.getExecution());
+        try {
+            createRevision(delegateTask.getExecution());
+        } catch (IOException e) {
+            handleException(delegateTask.getExecution(), ExceptionSource.TASK, e);
+        }
+
     }
 
-    private void createRevision(DelegateExecution execution) {
+    private void createRevision(DelegateExecution execution) throws IOException {
         String submissionId = formSubmissionService.createRevision(String.valueOf(execution.getVariables().get("formUrl")));
         execution.setVariable("formUrl", getUrl(execution) + "/" + submissionId);
     }
@@ -42,5 +53,7 @@ public class FormSubmissionListener implements ExecutionListener, TaskListener {
     private String getUrl(DelegateExecution execution){
         return StringUtils.substringBeforeLast(String.valueOf(execution.getVariables().get("formUrl")),"/");
     }
+
+
 
 }