+++ forms-flow-bpm/src/main/java/org/camunda/bpm/extension/hooks/listeners/task/NotifyListener.java
@@ -4,18 +4,20 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.nimbusds.oauth2.sdk.util.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
+import org.camunda.bpm.engine.ProcessEngineException;
 import org.camunda.bpm.engine.delegate.DelegateExecution;
 import org.camunda.bpm.engine.delegate.DelegateTask;
 import org.camunda.bpm.engine.delegate.Expression;
 import org.camunda.bpm.engine.delegate.TaskListener;
+import org.camunda.bpm.extension.hooks.listeners.BaseListener;
 import org.camunda.bpm.extension.hooks.services.IMessageEvent;
 import org.springframework.stereotype.Component;
 
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.logging.Level;
 import java.util.logging.Logger;
 
 /**
@@ -25,7 +27,7 @@ import java.util.logging.Logger;
  * @author yichun.zhao@aot-technologies.com, sumathi.thirumani@aot-technologies.com
  */
 @Component
-public class NotifyListener implements TaskListener, IMessageEvent {
+public class NotifyListener extends BaseListener implements TaskListener, IMessageEvent {
 
     private static final Logger LOGGER = Logger.getLogger(NotifyListener.class.getName());
 
@@ -44,11 +46,15 @@ public class NotifyListener implements TaskListener, IMessageEvent {
         List<String> toEmails =  new ArrayList<>();
         if(!"Y".equals(getGroupsOnly(delegateTask.getExecution()))) {
             toEmails.addAll(getEmailsOfUnassignedTask(delegateTask));
-
         }
-
-        if (CollectionUtils.isNotEmpty(getEmailGroups(delegateTask.getExecution()))) {
-            for (String entry : getEmailGroups(delegateTask.getExecution())) {
+        List<String> emailGroups = null;
+        try {
+            emailGroups = getEmailGroups(delegateTask.getExecution());
+        } catch (IOException e) {
+           handleException(delegateTask.getExecution(), ExceptionSource.TASK, e);
+        }
+        if (CollectionUtils.isNotEmpty(emailGroups)) {
+            for (String entry : emailGroups) {
                 toEmails.addAll(getEmailsForGroup(delegateTask.getExecution(), entry));
             }
         }
@@ -90,16 +96,12 @@ public class NotifyListener implements TaskListener, IMessageEvent {
         return String.valueOf(this.messageId.getValue(delegateExecution));
     }
 
-    private List<String> getEmailGroups(DelegateExecution delegateExecution){
+    private List<String> getEmailGroups(DelegateExecution delegateExecution) throws JsonProcessingException {
         List<String> emailGroups = new ArrayList<>();
-        try {
-            if(this.emailGroups != null &&
-                    StringUtils.isNotBlank(String.valueOf(this.emailGroups.getValue(delegateExecution)))) {
-                emailGroups = this.emailGroups != null && this.emailGroups.getValue(delegateExecution) != null ?
-                        getObjectMapper().readValue(String.valueOf(this.emailGroups.getValue(delegateExecution)), List.class) : null;
-            }
-        } catch (JsonProcessingException e) {
-            LOGGER.log(Level.SEVERE, "Exception occured in reading additionalEmailGroups" , e);
+        if(this.emailGroups != null &&
+                StringUtils.isNotBlank(String.valueOf(this.emailGroups.getValue(delegateExecution)))) {
+            emailGroups = this.emailGroups != null && this.emailGroups.getValue(delegateExecution) != null ?
+                    getObjectMapper().readValue(String.valueOf(this.emailGroups.getValue(delegateExecution)), List.class) : null;
         }
         return  emailGroups;
     }