+++ forms-flow-web/src/components/Insights/index.js
@@ -2,8 +2,6 @@ import React from "react";
 import Insights from "./Insights";
 import "./Insights.scss";
 
-const InsightsPage = React.memo(() => {
+export default React.memo(() => {
   return <Insights />;
 });
-
-export default InsightsPage;