+++ CHANGELOG.md
@@ -1,6 +1,63 @@
 # Changelog for formsflow.ai
 Mark  items as `Added`, `Changed`, `Fixed`, `Removed`, `Untested Features`, `Upcoming Features`
 
+## 4.0.1 - 2021-07-13
+
+`Added`
+
+**forms-flow-api**
+* Support for allowing CORS with multiple comma-separated origins.
+* Added authorization on the application details page based on user roles.
+
+**forms-flow-bpm**
+* Added new workflows - `One-Step Approval Process` and `Two-Step Approval Process`.
+
+**forms-flow-forms**
+* Added new forms- `Create New Business License Application` and `Freedom of Information and Protection of Privacy`.
+
+**forms-flow-web**
+* Show/hide Application Menu based on keycloak group.
+* Show/hide View Submissions button in form webpage based on keycloak group.
+* Add 404 page.
+* Add 403 page.
+
+`Fixed`
+
+**forms-flow-analytics**
+* Fixed the failing installation of the analytics component.
+
+**forms-flow-api**
+* Fix application details API not displaying values to client users.
+* Fixed the issue of not creating applications when called from the BPM side with process-instance-id.
+
+**forms-flow-bpm**
+* Fix done for authentication issue with Keycloak in the Keycloak configuration.
+* Fix done for single result query fetching multiple record's during formio REST call.
+
+**forms-flow-web**
+* Resolve Last Modified column on the client Application page is not working. 
+* Fix Application search icons breaking. 
+* Resolve Mime type issue in the webpage. 
+
+`Modified`
+
+**forms-flow-bpm**
+* formio token generation cycle reduced from 24 hours to 3.50 Hours.
+* Modified checked exception's on Listener services to Runtime exception.
+* Modified application logging package to Camunda base package level.
+
+**forms-flow-web**
+* Modify WebSocket implementation to support reconnection in Task Menu.
+* Footer was modified to display formsflow.ai with the version number.
+
+`Generic Changes`
+* Improved the README to document supported version for Keycloak.
+* Updated [usage docs](./USAGE.md) with the latest form and workflow.
+* v1.0.7 release for `camunda-formio-tasklist-vue`,a Vue.js-based package for easy integration of formsflow.ai to existing projects. To know more details checkout [formsflow-ai-extension repository](https://github.com/AOT-Technologies/forms-flow-ai-extensions/tree/master/camunda-formio-tasklist-vue) 
+
+`Known Issues`
+* Consoles related to http://localhost:3001/current Api Failing. The console messages can be ignored. Please refer to [Issue-#106](https://github.com/AOT-Technologies/forms-flow-ai/issues/106) for more details.
+
 ## 4.0.0 - 2021-06-11
 
 `Added`
@@ -84,7 +141,6 @@ Mark  items as `Added`, `Changed`, `Fixed`, `Removed`, `Untested Features`, `Upc
 `Known Issues`
 * Custom component (Text Area with analytics) not retaining the value after submission
 * Cosmetic changes to show success message after loading is completed
-
 
 ## 2.0.1 - 2020-07-27
 `Added`